requirements.txt
backports.functools-lru-cache==1.6.1
certifi==2019.11.28
cloudpickle==1.3.0
cycler==0.10.0
decorator==4.4.2
enum34==1.1.10
face-alignment==1.1.1
kiwisolver==1.1.0
matplotlib==2.2.5
networkx==2.2
numpy==1.16.6
opencv-python==3.3.0.10
pandas==0.24.2
Pillow==6.2.2
pyparsing==2.4.7
python-dateutil==2.8.1
pytz==2020.1
PyWavelets==1.0.3
scikit-image==0.14.5
scipy==1.2.3
six==1.15.0
subprocess32==3.5.4
torch==1.0.0
torchvision==0.2.0
tqdm==4.50.0
trimesh==3.8.10
